% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{addAverageFoldChangeColumn}
\alias{addAverageFoldChangeColumn}
\title{add Average Fold Change statistics to enrichment Results}
\usage{
addAverageFoldChangeColumn(enrichmentResults, comparisonExperiment)
}
\description{
add Average Fold Change statistics to enrichment Results
}
