% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{check_genesets_left}
\alias{check_genesets_left}
\title{compute_unmapped_prop}
\usage{
check_genesets_left(
  genesets_items_final,
  genesets_table_final,
  msig_table,
  max_prop_unmapped,
  sets_removed
)
}
\arguments{
\item{geneset_genes}{table with gene sets and proteins/genes mapped to them}

\item{genesets_table}{table with gene sets and their infos}
}
\description{
compute_unmapped_prop
}
