% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{compute_unmapped_prop}
\alias{compute_unmapped_prop}
\title{compute_unmapped_prop}
\usage{
compute_unmapped_prop(geneset_genes)
}
\arguments{
\item{geneset_genes}{table with gene sets and proteins/genes mapped to them}
}
\description{
compute_unmapped_prop
}
