% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{create_gene_set_table}
\alias{create_gene_set_table}
\title{create_gene_set_table}
\usage{
create_gene_set_table(msigdb_table)
}
\arguments{
\item{msigdb_table}{tibble with data imported with \code{msigdb_xml_to_table()}}
}
\description{
create_gene_set_table
}
