% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_reactome_sets.R
\name{create_reactome_sets}
\alias{create_reactome_sets}
\title{create_reactome_set}
\usage{
create_reactome_sets(species)
}
\arguments{
\item{species}{string specifying the species to create the gene sets for}
}
\description{
create_reactome_set
}
