% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{createColumnData}
\alias{createColumnData}
\title{Creates the Column Data (experimental design) for a given experimental Comparison}
\usage{
createColumnData(
  condition1,
  condition2,
  conditionRunIdMapping,
  conditionCol = "Condition"
)
}
\description{
Creates the Column Data (experimental design) for a given experimental Comparison
}
