% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareIntensities.R
\name{createComparisonSummarisedExperiment}
\alias{createComparisonSummarisedExperiment}
\title{Create a SummarizedExperiment object from wide-format protein intensity data}
\usage{
createComparisonSummarisedExperiment(
  wideIntensities,
  leftCondition,
  rightCondition,
  replicateConditionMapping,
  idCol,
  conditionColname,
  replicateColname
)
}
\arguments{
\item{wideIntensities}{A data frame containing protein intensity values in wide format as created from \code{makeWideIntensitiesDf}}
}
\value{
A SummarizedExperiment object containing the protein intensity values
}
\description{
Given a data frame of protein intensity values in wide format, this function
converts it to a SummarizedExperiment object, which is a Bioconductor class
for storing and manipulating genomic data. The resulting object can be used
for downstream analysis and visualization.
}
