% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{createEnrichmentDatasetParameters}
\alias{createEnrichmentDatasetParameters}
\title{createEnrichmentDatasetParameters}
\usage{
createEnrichmentDatasetParameters(dataset_id, version, fields)
}
\description{
createEnrichmentDatasetParameters
}
