% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.getConditionComparisons}
\alias{datasets.getConditionComparisons}
\title{Gets the condition comparisons provided by the dataset parameters}
\usage{
datasets.getConditionComparisons(datasetParameters)
}
\arguments{
\item{datasetParameters}{a jsonlite representation of the dataset parameters}
}
\value{
a data frame representing the condition comparisons
}
\description{
Gets the condition comparisons provided by the dataset parameters
}
