% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.getGeneSetsInfo}
\alias{datasets.getGeneSetsInfo}
\title{Gets the gene sets information provided by the dataset parameters}
\usage{
datasets.getGeneSetsInfo(datasetParameters)
}
\arguments{
\item{datasetParameters}{a jsonlite representation of the dataset parameters}
}
\value{
a list representing the gene sets information
}
\description{
Gets the gene sets information provided by the dataset parameters
}
