% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{define_genesets_to_remove}
\alias{define_genesets_to_remove}
\title{define_genesets_to_remove}
\usage{
define_genesets_to_remove(genesets_table, max_prop_unmapped)
}
\arguments{
\item{genesets_table}{Table with gene sets and their information}

\item{max_prop_unmapped}{unmapped proportion to use for filtering gene sets}
}
\description{
define_genesets_to_remove
}
\keyword{internal}
