% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichComparisonExperiment}
\alias{enrichComparisonExperiment}
\title{Automates Enrichment Analyses on Summarized Objects for a library of gene sets}
\usage{
enrichComparisonExperiment(
  comparisonExperiment,
  gmtFolder,
  method,
  perm = perm
)
}
\arguments{
\item{gmtFolder}{a folder containing gene set libraries which need to be referenced}

\item{method}{the method enrichment browser should use for set based enrichment}

\item{comparison}{A summarized experiment object containing only two experimental groups}
}
\value{
enrichment An enrichment table provided by EnrichmentBrowserPackage
}
\description{
Automates Enrichment Analyses on Summarized Objects for a library of gene sets
}
