% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{enrichComparisons}
\alias{enrichComparisons}
\title{Orchestrates the enrichment workflow step for a comparisonExperiment List}
\usage{
enrichComparisons(
  comparisonExperimentsList,
  gmtFolder,
  method = "camera",
  perms = 100
)
}
\description{
Orchestrates the enrichment workflow step for a comparisonExperiment List
}
