% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{extract_one_entry_uniprot_kb}
\alias{extract_one_entry_uniprot_kb}
\title{extract_one_entry_uniprot_kb}
\usage{
extract_one_entry_uniprot_kb(map_one_gene)
}
\description{
extract_one_entry_uniprot_kb
}
\keyword{internal}
