% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{filterIntensitiesByProteinViz}
\alias{filterIntensitiesByProteinViz}
\title{Filters protein intensities by the proteins which the limma statistics
were calculated upon.}
\usage{
filterIntensitiesByProteinViz(proteinViz, proteinIntensitiesWide)
}
\description{
Filters protein intensities by the proteins which the limma statistics
were calculated upon.
}
