% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{filterProteinIntensityColumnsbyComparison}
\alias{filterProteinIntensityColumnsbyComparison}
\title{filter the protein intensity by columns for a comparison}
\usage{
filterProteinIntensityColumnsbyComparison(
  condition1,
  condition2,
  conditionRunIdMapping,
  assay_data
)
}
\description{
filter the protein intensity by columns for a comparison
}
