% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{generate_msig_sets}
\alias{generate_msig_sets}
\title{generate_msig_sets}
\usage{
generate_msig_sets(
  xml_filepath,
  msig_version,
  uniprot_table_path,
  max_prop_unmapped = 0.8
)
}
\arguments{
\item{xml_filepath}{file path to msigDB xml file}

\item{msig_version}{msig version}

\item{uniprot_table_path}{File path to table downloaded from uniprot}
}
\description{
generate_msig_sets
}
