% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_go_sets.R
\name{get_sets_category}
\alias{get_sets_category}
\title{Get GO sets category}
\usage{
get_sets_category(obo_file)
}
\arguments{
\item{obo_file}{obo file imported with \code{get_ontology}}
}
\description{
Get GO sets category
}
