% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_go_sets.R
\name{get_sets_information}
\alias{get_sets_information}
\title{Get sets information}
\usage{
get_sets_information()
}
\arguments{
\item{go_url}{url to gaf file}
}
\description{
Get sets information
}
