% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{getConditionComparisonMapping}
\alias{getConditionComparisonMapping}
\title{Get mapping from comparison strings to up and down conditions}
\usage{
getConditionComparisonMapping(proteinViz)
}
\arguments{
\item{proteinViz}{a table produced by md quant workflows with DE statistics}
}
\value{
a dataframe containing the up and down conditions for each comparison
}
\description{
Get mapping from comparison strings to up and down conditions
}
