% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{getConditionComparisons,EnrichmentDatasetParameters-method}
\alias{getConditionComparisons,EnrichmentDatasetParameters-method}
\title{getConditionComparisons method}
\usage{
\S4method{getConditionComparisons}{EnrichmentDatasetParameters}(params)
}
\description{
getConditionComparisons method
}
