% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{getConditionRunIdMapping}
\alias{getConditionRunIdMapping}
\title{Returns a dataframe matching Run Id's to Conditions}
\usage{
getConditionRunIdMapping(proteinIntensitiesLong)
}
\description{
Returns a dataframe matching Run Id's to Conditions
}
