% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnrichmentComparison.R
\name{getGeneSetList}
\alias{getGeneSetList}
\title{Automates Enrichment Analyses on Summarized Objects for a library of gene sets}
\usage{
getGeneSetList(species, geneSet, version)
}
\arguments{
\item{species}{species, e.g. Human, Mouse}

\item{geneSet}{name of gene set: Reactome, MolecularProcess etc..}
}
\description{
Automates Enrichment Analyses on Summarized Objects for a library of gene sets
}
