% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{getGeneSets,EnrichmentDatasetParameters-method}
\alias{getGeneSets,EnrichmentDatasetParameters-method}
\title{getGeneSets method}
\usage{
\S4method{getGeneSets}{EnrichmentDatasetParameters}(params)
}
\description{
getGeneSets method
}
