% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{getIntensities,IntensitiesTable-method}
\alias{getIntensities,IntensitiesTable-method}
\title{Get Intensities data from IntensitiesTable class}
\usage{
\S4method{getIntensities}{IntensitiesTable}(dataset)
}
\description{
Get Intensities data from IntensitiesTable class
}
