% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-mapping.R
\name{getSharedColNamesAnnotations}
\alias{getSharedColNamesAnnotations}
\title{Shared column names between annotation sets}
\usage{
getSharedColNamesAnnotations()
}
\description{
Shared column names between annotation sets
}
\keyword{internal}
