% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{getSpecies,EnrichmentDatasetParameters-method}
\alias{getSpecies,EnrichmentDatasetParameters-method}
\title{getSpecies method}
\usage{
\S4method{getSpecies}{EnrichmentDatasetParameters}(params)
}
\description{
getSpecies method
}
