% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{listcomparisonExperimentsList}
\alias{listcomparisonExperimentsList}
\title{list all the binary experiments comparisons used for enrichment}
\usage{
listcomparisonExperimentsList(mdQuantExports, conditionComparisonMapping)
}
\arguments{
\item{mdQuantExports}{a named list of mass dyanmics quantitation objects produced by loadMDQuantOutput}
}
\value{
a list of summarized experiments containing protein de expression intensities and statistics
}
\description{
list all the binary experiments comparisons used for enrichment
}
