% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{loadMDQuantOutput}
\alias{loadMDQuantOutput}
\title{Loads the protein quantitation output from an md workflow step}
\usage{
loadMDQuantOutput(uploadFolder)
}
\arguments{
\item{uploadFolder}{a folder containing mass dynamics quantitation output}
}
\value{
a named list of mass dyanmics quantitation objects.
}
\description{
Loads the protein quantitation output from an md workflow step
}
