% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_intensities.R
\name{makeWideIntensitiesDf}
\alias{makeWideIntensitiesDf}
\title{Convert protein intensity data from long to wide format}
\usage{
makeWideIntensitiesDf(
  proteinIntensities,
  groupByCol = "ProteinIds",
  intensityCol = "NormalisedIntensity",
  replicateCol = "replicate"
)
}
\arguments{
\item{proteinIntensities}{A data frame containing protein intensity in long format as imported from "Protein_intensity.tsv"}

\item{groupByCol}{The name of the column in \code{proteinIntensities} that specifies the grouping variable
to be used for creating the wide format. Default is "ProteinIds".}

\item{intensityCol}{The name of the column in \code{proteinIntensities} that contains the protein intensity values.
Default is "NormalisedIntensity".}

\item{replicateCol}{The name of the column in \code{proteinIntensities} that specifies the replicates.
Default is "replicate".}
}
\value{
A data frame containing the protein intensity values in wide format
}
\description{
Given a data frame of protein intensity values in long format, this function
converts it to wide format, with one row per protein and one column per
replicate. The wide format is useful for downstream analysis and visualization.
}
