% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{manualRunCameraEnrichment}
\alias{manualRunCameraEnrichment}
\title{Wrapper for CAMERA}
\usage{
manualRunCameraEnrichment(comparisonExperiment, geneSets)
}
\arguments{
\item{comparisonExperiment}{A summarized experiment object containing only two experimental groups}

\item{geneSets}{gene sets list to run CAMERA}
}
\description{
Wrapper for CAMERA
}
