% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{map_items_to_uniprot}
\alias{map_items_to_uniprot}
\title{map_items_to_uniprot}
\usage{
map_items_to_uniprot(msigdb_table, uniprot_table_path)
}
\arguments{
\item{msigdb_table}{tibble with data imported with \code{msigdb_xml_to_table()}}

\item{uniprot_table_path}{File path to table downloaded from uniprot}
}
\description{
map_items_to_uniprot
}
