% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-mapping.R
\name{map_sets_to_file_human}
\alias{map_sets_to_file_human}
\title{Map sets to human R files}
\usage{
map_sets_to_file_human(set)
}
\arguments{
\item{set}{Name of the annotation set}
}
\description{
Map sets to human R files
}
