% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-mapping.R
\name{map_sets_to_file_mouse}
\alias{map_sets_to_file_mouse}
\title{Map sets to mouse R files}
\usage{
map_sets_to_file_mouse(set)
}
\arguments{
\item{set}{Name of the annotation set}
}
\description{
Map sets to mouse R files
}
