% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-mapping.R
\name{map_sets_to_file_yeast}
\alias{map_sets_to_file_yeast}
\title{Map sets to yeast R files}
\usage{
map_sets_to_file_yeast(set)
}
\arguments{
\item{set}{Name of the annotation set}
}
\description{
Map sets to yeast R files
}
