% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{map_symbols_to_uniprotids}
\alias{map_symbols_to_uniprotids}
\title{map_symbols_to_uniprotids}
\usage{
map_symbols_to_uniprotids(items, uniprot_table)
}
\arguments{
\item{items}{list of genes}

\item{uniprot_table}{Table downloaded from uniprot}
}
\description{
Takes a list of genes symbols, the UNIPROT table containing both the "Entry" (uniprot id) and the 'Gene Names (primary)' columns
and map gene symbols to uniprotIDs.
}
