% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{md2ProteinSummarizedExperiment}
\alias{md2ProteinSummarizedExperiment}
\title{This function generates the protein summarized experiment from the processed md artifacts}
\usage{
md2ProteinSummarizedExperiment(
  rowDataStatistics,
  proteinIntensitiesComparison,
  columnData,
  up.condition,
  down.condition,
  by = "Protein"
)
}
\description{
This function generates the protein summarized experiment from the processed md artifacts
}
