% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{msigdb_xml_to_table}
\alias{msigdb_xml_to_table}
\title{generate_msig_sets}
\usage{
msigdb_xml_to_table(xml_filepath)
}
\arguments{
\item{xml_filepath}{file path to msigDB xml file}
}
\description{
generate_msig_sets
}
