% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{perform_comparison_enrichment}
\alias{perform_comparison_enrichment}
\title{Automates Enrichment Analyses on Summarized Objects for a library of gene sets}
\usage{
perform_comparison_enrichment(comparison, gmt_folder)
}
\arguments{
\item{comparison}{A summarized experiment object containing only two experimental groups}
}
\value{
enrichment An enrichment table provided by EnrichmentBrowserPackage
}
\description{
Automates Enrichment Analyses on Summarized Objects for a library of gene sets
}
\examples{
sdgdfsgsf
}
