% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_intensities.R
\name{prepareComparisonExperiment}
\alias{prepareComparisonExperiment}
\title{Prepare comparison experiment}
\usage{
prepareComparisonExperiment(
  intensitiesDf,
  metadataDf,
  conditionComparison,
  conditionCol,
  replicateCol,
  featureIdCol,
  intensityCol,
  imputedCol
)
}
\arguments{
\item{intensitiesDf}{A data.frame containing intensity values for features.}

\item{metadataDf}{A data.frame containing metadata for each feature.}

\item{conditionComparison}{list containing levels of conditions to be compared.}

\item{conditionCol}{A character string specifying the column name in intensitiesDf that contains the condition information.}

\item{replicateCol}{A character string specifying the column name in intensitiesDf that contains the replicate information.}

\item{featureIdCol}{A character string specifying the column name in intensitiesDf and metadataDf that contains the feature identifier information.}

\item{intensityCol}{A character string specifying the column name in intensitiesDf that contains the intensities.}

\item{imputedCol}{A character string specifying the column name in intensitiesDf that contains the imputed value information.}

\item{replicateConditionMapping}{A list specifying the mapping between replicates and conditions.}
}
\value{
a SummarizedExperiment object containing the intensities, column and row data to perform gene set testing.
}
\description{
This function prepares a comparison experiment by organizing input data into a standardized format.
}
