% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareIntensities.R
\name{prepareRowData}
\alias{prepareRowData}
\title{Prepare row data}
\usage{
prepareRowData(
  intAssay,
  idCol,
  replicateMapping,
  leftCondition,
  rightCondition
)
}
\description{
Prepare row data
}
