% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{rename_genes_protein_mapping_columns}
\alias{rename_genes_protein_mapping_columns}
\title{rename_genes_protein_mapping}
\usage{
rename_genes_protein_mapping_columns(genes_protein_mapping)
}
\arguments{
\item{genes_protein_mapping}{Table with gene sets, gene names, and uniprot id mapping}
}
\description{
rename_genes_protein_mapping
}
