% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{rename_genesets_items_columns}
\alias{rename_genesets_items_columns}
\title{rename_genesets_items}
\usage{
rename_genesets_items_columns(genesets_items)
}
\arguments{
\item{genesets_items}{Table with mapping between gene sets and their items}
}
\description{
rename_genesets_items
}
