% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-create_msigb_sets.R
\name{rename_genesets_table_columns}
\alias{rename_genesets_table_columns}
\title{rename_geneset_table}
\usage{
rename_genesets_table_columns(genesets_table)
}
\arguments{
\item{genesets_table}{Table with gene sets and their infos}
}
\description{
rename_geneset_table
}
