% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{runCamera}
\alias{runCamera}
\title{Wrapper for CAMERA}
\usage{
runCamera(comparisonExperiment, geneSets, preRanked = FALSE)
}
\arguments{
\item{comparisonExperiment}{A summarized experiment object containing only two experimental groups}

\item{geneSets}{gene sets list to run CAMERA}
}
\description{
Wrapper for CAMERA
}
