% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnrichmentComparison.R
\name{runEnrichmentForAnnotationSet}
\alias{runEnrichmentForAnnotationSet}
\title{Automates Enrichment Analyses for one dataaset}
\usage{
runEnrichmentForAnnotationSet(intensitiesDf, metadataDf, datasetParams)
}
\arguments{
\item{intensitiesDf}{List of three list each containing a pair of names and tables for: protein metadata,
intensities, condition comparison}

\item{metadataDf}{Features metadata}

\item{datasetParams}{list of metadata passed by the user}
}
\description{
Automates Enrichment Analyses for one dataaset
}
