% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnrichmentComparison.R
\name{runEnrichmentOneAnnotation}
\alias{runEnrichmentOneAnnotation}
\title{Automates Enrichment Analyses on Summarized Objects for a library of gene sets}
\usage{
runEnrichmentOneAnnotation(
  intensityDataset,
  conditionComparisons,
  geneSetsList,
  database,
  method = "camera",
  conditionCol = "condition",
  replicateCol = "replicate",
  featureIdCol = "GroupId",
  intensityCol = "NormalisedIntensity",
  imputedCol = "Imputed"
)
}
\arguments{
\item{method}{the method enrichment browser should use for set based enrichment}

\item{comparison}{A summarized experiment object containing only two experimental groups}

\item{gmtFolder}{a folder containing gene set libraries which need to be referenced}
}
\value{
enrichment An enrichment table provided by EnrichmentBrowserPackage
}
\description{
Automates Enrichment Analyses on Summarized Objects for a library of gene sets
}
