% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{runEnrichmentWorkflowStep}
\alias{runEnrichmentWorkflowStep}
\title{Orchestrates the enrichment workflow step for MD input/output data}
\usage{
runEnrichmentWorkflowStep(
  uploadFolder,
  gmtFolder,
  outputFolder,
  by = "Protein",
  method = "camera",
  perm
)
}
\description{
Orchestrates the enrichment workflow step for MD input/output data
}
