% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{runGeneSetTest}
\alias{runGeneSetTest}
\title{Wrapper run generic gene set tests}
\usage{
runGeneSetTest(comparisonExperiment, geneSets, method)
}
\arguments{
\item{comparisonExperiment}{A summarized experiment object containing only two experimental groups}

\item{geneSets}{gene sets list}
}
\description{
Wrapper run generic gene set tests
}
