% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-species_mapping.R
\name{species_short_to_long}
\alias{species_short_to_long}
\title{Species short to long names}
\usage{
species_short_to_long(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy code}
}
\description{
Species short to long names
}
