% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-species_mapping.R
\name{species_to_taxonomy}
\alias{species_to_taxonomy}
\title{Species to taxonomy}
\usage{
species_to_taxonomy(species)
}
\arguments{
\item{species}{taxonomy code}
}
\description{
Species to taxonomy
}
