% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser_md.R
\name{spreadProteinIntensities}
\alias{spreadProteinIntensities}
\title{Long to wide conversion of protein Intensities}
\usage{
spreadProteinIntensities(proteinIntensitiesLong)
}
\arguments{
\item{proteinIntensitiesLong}{a long form protein intensities object produced by the md quant workflow}
}
\value{
a wide dataframe containing all the processed protein intensity information
}
\description{
Long to wide conversion of protein Intensities
}
