% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{subsetCategory,AnnotationSets-method}
\alias{subsetCategory,AnnotationSets-method}
\title{subsetCategory in gene sets method}
\usage{
\S4method{subsetCategory}{AnnotationSets}(.Object, category)
}
\description{
subsetCategory in gene sets method
}
