% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-methods.R
\name{subsetCategory,AnnotationSetsMsigDB-method}
\alias{subsetCategory,AnnotationSetsMsigDB-method}
\title{subsetCategory in gene sets method}
\usage{
\S4method{subsetCategory}{AnnotationSetsMsigDB}(.Object, category)
}
\description{
subsetCategory in gene sets method
}
