% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-species_mapping.R
\name{taxonomy_to_long_species}
\alias{taxonomy_to_long_species}
\title{Taxonomy to long species name}
\usage{
taxonomy_to_long_species(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy code}
}
\description{
Taxonomy to long species name
}
