% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-species_mapping.R
\name{taxonomy_to_short_species}
\alias{taxonomy_to_short_species}
\title{Taxonomy to short species name}
\usage{
taxonomy_to_short_species(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy code}
}
\description{
Taxonomy to short species name
}
