% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets-write_output.R
\name{write_set_info_reactome}
\alias{write_set_info_reactome}
\title{Write Set infos for Reactome}
\usage{
write_set_info_reactome(reactome_sets, outfolder = "../inst/extdata/")
}
\description{
Write Set infos for Reactome
}
