% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEnrichmentComparison.R
\name{runEnrichmentForAnnotationSet}
\alias{runEnrichmentForAnnotationSet}
\title{Automates Enrichment Analyses on Summarized Objects for a library of gene sets}
\usage{
runEnrichmentForAnnotationSet(datasetParams)
}
\arguments{
\item{datasetsTables}{List of three list each containing a pair of names and tables for: protein metadata,
intensities, condition comparison}

\item{metadata}{list of metadata passed by the user}
}
\description{
Automates Enrichment Analyses on Summarized Objects for a library of gene sets
}
